+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added `DOMNode.displayed_children` https://github.com/Textualize/textual/pull/6070
 - Added `TextArea.UserInsert` message https://github.com/Textualize/textual/pull/6070
 - Added `TextArea.hide_suggestion_on_blur` boolean https://github.com/Textualize/textual/pull/6070
+- Added `DataTable.get_row_key` https://github.com/Textualize/textual/pull/6081
 
 ### Changed